pyproject.toml
[tool.poetry]
name = "openai-content"
version = "0.1.0"
description = ""
authors = ["Hasnain Hasib <hasibjoy332@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.112.2"
sqlalchemy = "^2.0.33"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
openai = "0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
